cart更新

+++ cart/src/main/java/xmu/oomall/controller/CartControllerImpl.java
@@ -10,13 +10,13 @@ import xmu.oomall.service.CartService;
 import xmu.oomall.service.GoodsService;
 
 import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
  * @author liznsalt
  */
 @RestController
-@RequestMapping("/cartService")
 public class CartControllerImpl {
 
     @Autowired
@@ -37,6 +37,15 @@ public class CartControllerImpl {
         Object object = goodsService.getProductById(id);
         Product product = JacksonUtil.parseObject(JacksonUtil.toJson(object),
                 "data", Product.class);
+        if (product == null) {
+            Integer errno = JacksonUtil.parseInteger(JacksonUtil.toJson(object),
+                    "errno");
+            if (errno != null && errno.equals(0)) {
+                return new Product();
+            } else {
+                return null;        // connect error
+            }
+        }
         return product;
     }
 
@@ -48,25 +57,34 @@ public class CartControllerImpl {
                 || cartItem.getNumber() < 0) {
             return false;
         }
+        Product product = findProductById(cartItem.getProductId());
+        if (product == null || product.getId() == null) {
+            return false;
+        }
         return true;
     }
 
     // 内部接口
-    @Deprecated
     @GetMapping("/cartItem/{userId}")
     public Object userCart(@PathVariable Integer userId) {
         if (userId == null || userId < 0) {
             return CommonResult.unLogin();
         }
         List<MallCartItem> cartItems = cartService.list(userId);
+        List<MallCartItem> resultCartItems = new ArrayList<>();
         for (MallCartItem mallCartItem : cartItems) {
             Product product = findProductById(mallCartItem.getProductId());
             if (product == null) {
                 return CommonResult.serious();
             }
-            mallCartItem.setProduct(product);
+            if (product.getId() == null) {
+                cartService.delete(mallCartItem.getId());
+            } else {
+                mallCartItem.setProduct(product);
+                resultCartItems.add(mallCartItem);
+            }
         }
-        return CommonResult.success(cartItems);
+        return CommonResult.success(resultCartItems);
     }
 
 
@@ -80,14 +98,21 @@ public class CartControllerImpl {
         }
 
         List<MallCartItem> cartItems = cartService.list(userId);
+        List<MallCartItem> resultCartItems = new ArrayList<>();
+        // filter items whose product is wrong
         for (MallCartItem mallCartItem : cartItems) {
             Product product = findProductById(mallCartItem.getProductId());
             if (product == null) {
                 return CommonResult.serious();
             }
-            mallCartItem.setProduct(product);
+            if (product.getId() == null) {
+                cartService.delete(mallCartItem.getId());
+            } else {
+                mallCartItem.setProduct(product);
+                resultCartItems.add(mallCartItem);
+            }
         }
-        return CommonResult.success(cartItems);
+        return CommonResult.success(resultCartItems);
     }
 
     @GetMapping("/cartItems/{id}")
@@ -110,12 +135,16 @@ public class CartControllerImpl {
         if (product == null) {
             return CommonResult.serious();
         }
+        if (product.getId() == null) {
+            cartService.delete(id);
+            return CommonResult.failed();
+        }
         cartItem.setProduct(product);
         return CommonResult.success(cartItem);
     }
 
     @PostMapping("/cartItems")
-    public Object add(MallCartItem cart, HttpServletRequest request) {
+    public Object add(@RequestBody MallCartItem cart, HttpServletRequest request) {
         Integer userId = getUserId(request);
         if (userId == null || userId < 0) {
             return CommonResult.unLogin();
@@ -160,21 +189,27 @@ public class CartControllerImpl {
     }
 
     @PutMapping("/cartItems/{id}")
-    public Object update(@PathVariable Integer id, MallCartItem cart, HttpServletRequest request) {
+    public Object update(@PathVariable Integer id, @RequestBody MallCartItem cart,
+                         HttpServletRequest request) {
         Integer userId = getUserId(request);
         if (userId == null || userId < 0) {
             return CommonResult.unLogin();
         }
-
         // 参数校验
         if (id == null || id < 0) {
             return CommonResult.badArgumentValue();
+        }
+        MallCartItem dbCartItem = cartService.findCartItemById(id);
+        if (dbCartItem == null) {
+            return CommonResult.badArgumentValue();
+        }
+        if (!userId.equals(dbCartItem.getUserId())) {
+            return CommonResult.unauthorized();
         }
         cart.setId(id);
         if (!isValid(cart)) {
             return CommonResult.badArgumentValue();
         }
-
         MallCartItem cartItem = cartService.update(userId, cart);
         if (cartItem == null || cartItem.getId() == null) {
             return CommonResult.updatedDataFailed();
@@ -195,6 +230,14 @@ public class CartControllerImpl {
             return CommonResult.badArgumentValue();
         }
 
+        MallCartItem cartItem = cartService.findCartItemById(id);
+        if (cartItem == null) {
+            return CommonResult.failed();
+        }
+        if (!userId.equals(cartItem.getUserId())) {
+            return CommonResult.unauthorized();
+        }
+
         boolean ok = cartService.delete(id);
         if (ok) {
             return CommonResult.success(null);
@@ -204,7 +247,7 @@ public class CartControllerImpl {
     }
 
     @PostMapping("/fastAddCartItems")
-    public Object fastAddCartItems(MallCartItem cartItem,
+    public Object fastAddCartItems(@RequestBody MallCartItem cartItem,
                                    HttpServletRequest request) {
         Integer userId = getUserId(request);
         if (userId == null) {
